+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Change: Update copyright date (2024)
 * New: Add `.gitignore` file
 * Change: Reduce number of 'Tags' from `readme.txt`
+* Unit tests:
+    * Change: In bootstrap, store path to plugin directory in a constant
 
 ## 1.4.3 _(2023-04-29)_
 * Change: Note compatibility through WP 6.3+