+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * New: Add `.gitignore` file
 * Change: Reduce number of 'Tags' from `readme.txt`
 * Unit tests:
+    * Allow tests to run against current versions of WordPress
+    * New: Add `composer.json` for PHPUnit Polyfill dependency
     * Change: In bootstrap, store path to plugin directory in a constant
 
 ## 1.4.3 _(2023-04-29)_