+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * New: Add `.gitignore` file
 * Change: Reduce number of 'Tags' from `readme.txt`
 * Unit tests:
+    * Hardening: Prevent direct web access to `bootstrap.php`
     * Allow tests to run against current versions of WordPress
     * New: Add `composer.json` for PHPUnit Polyfill dependency
     * Change: In bootstrap, store path to plugin directory in a constant